requirements.txt

requests
beautifulsoup4

pandas

fastapi
uvicorn[standard]

streamlit
sqlalchemy
databases

passlib[bcrypt]
PyJWT
python-multipart